Fix: Week06 All Requirements Done
