+++ src/main/java/jwp/WebServerLauncher.java
@@ -1,0 +1,11 @@
+package jwp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WebServerLauncher {
+    public static void main(String[] args) throws Exception {
+        SpringApplication.run(WebServerLauncher.class, args);
+    }
+}
