+++ src/main/java/jwp/controller/Controller.java
@@ -1,11 +1,11 @@
-package jwp.controller;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.sql.SQLException;
-
-public interface Controller {
-    String execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, SQLException;
-}
+//package jwp.controller;
+//
+//import javax.servlet.ServletException;
+//import javax.servlet.http.HttpServletRequest;
+//import javax.servlet.http.HttpServletResponse;
+//import java.io.IOException;
+//import java.sql.SQLException;
+//
+//public interface Controller {
+//    String execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, SQLException;
+//}
